Improved lightning type

+++ src/main/java/pl/put/bulidinginfo/models/Room.java
@@ -4,9 +4,9 @@ public class Room extends Location {
     public Float area;
     public Float cube;
     public Float heating;
-    public Float lighting;
+    public Integer lighting;
 
-    public Room(Integer id, String name, Float area, Float cube, Float heating, Float lighting) {
+    public Room(Integer id, String name, Float area, Float cube, Float heating, Integer lighting) {
         super(id, name);
         this.area = area;
         this.type = Type.ROOM;
